move stuff around to allow for unit testing
