+++ lovebrew/consoles/cafe.py
@@ -1,8 +1,8 @@
 import shutil
 from pathlib import Path
 
-from console import Console
-from error import Error
+from lovebrew.console import Console
+from lovebrew.error import Error
 
 
 class Cafe(Console):
