+++ lovebrew/consoles/hac.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
-from console import Console
-from error import Error
+from lovebrew.console import Console
+from lovebrew.error import Error
 
 
 class Hac(Console):
