+++ pyproject.toml
@@ -1,0 +1,19 @@
+[project]
+name = "LOVEBrew"
+description = "A web application to build your LOVE Potion projects"
+license = "MIT"
+requires-python = ">=3.8"
+dependencies = [
+    "flask",
+    "waitress",
+    "pyopenssl",
+    "requests",
+    "filetype",
+    "semver",
+]
+dynamic = ["version"]
+
+[project.urls]
+Donate = "https://paypal.me/TurtleP"
+"Source Code" = "https://github.com/lovebrew/lovebrew-webserver/"
+"Issue Tracker" = "https://github.com/lovebrew/lovebrew-webserver/issues/"
