--- requirements.txt
@@ -1,6 +1,0 @@
-flask
-waitress
-pyopenssl
-requests
-filetype
-semver
